[447042] Stop sending Workspace related events in an async way
When the session is saving, it is waiting for a workspace event related
to the files it just saved and will identify those as "expected", hence
not requiring a resource reload. By sending these events in an async way
we are just adding more entropy in this regard, the session might
retrieve a completely unrelated event and act accordingly.

 
Bug: 447042
Change-Id: I868f807e45e8bc5e749b01469f1b94454b975f81
Signed-off-by: Cedric Brun <cedric.brun@obeo.fr>