[456350] Use a JavaExtensionManager to init the ModelAccessor
Instead of browsing through all the model and call "notifyNewMetamodel"
when an unknown one is found, uses a JavaExtensionManager to compute and
update the list of EPackages which are accessible through the classpath.

This avoids a complete traversal of the resourceset during
Session.open(). Also removes the  if (!getSemanticResources().isEmpty())
condition which was guarding the configuration of the interpreter
instance as it should not depend on the model content.
Bug: 456350
Change-Id: I20c2e96883d5535f61909db885e9d5080a9b225d