[466718] DRAFT -- Don't visit Mappings which can't be overriden
Implementing the strategy of determining upfront that a XXXMapping is
not overriden in any way in the current enabled layers and making the
iteration over the hiearchy to -- maybe -- update the actualMapping
reference not necessary.

This is done to avoid multiple evaluations of the semantic candidate
expressions for a given mapping and context.

Bug: 466718
Change-Id: I9639df9fd42dd0248135424907d616fcafabadb8
Signed-off-by: Cedric Brun <cedric.brun@obeo.fr>