[DRAFT] : stop recording TOUCH notifications
A significant time of processing is spent in the ChangeRecorder
aggregating the notifications. In the case of a  diagram having 2K
ddiagram elements and during a refresh which *will not cause any change
in the diagram model* about 10% of the time is spent in the
ChangeRecorder. In that case this is completely due to the fact that the
ChangeRecorder will *also* record TOUCH notifications.

This patchset changes the TransactionChangeRecorder so that it ignores
the TOUCH notifications, this has implications on any listener
(SessionEventBroker, ResourceSetListenerImpl) which *would expect* such
notifications.

Change-Id: I709ba5e78a24823e9e2bf1b35dd099b2c9c9c9b4
Signed-off-by: Cedric Brun <cedric.brun@obeo.fr>