[472811] Added registration of the EClass instance Class for workspace ecore.
When Sirius register an EPackage from the workspace it use the .ecore resource.
EClasses form the EPackage don't have instance Class associated when loaded
this way. So we use the .genmodel resource to manually map EClasses to their
instance Class. This way we can propose Java services for those EClasses.

Bug: 472811
Change-Id: I811e2f6f490f1252c8d6e3fa9e01835614cc1d36
Signed-off-by: Yvan Lussaud <yvan.lussaud@obeo.fr>