[DRAFT/POC 493353] Set of modifications to allow support of controlled representations
SessionResourceTracker semantic resource management:
Old behavior was to get the first root of a DSemanticDecorator's
resource.
Javadoc will be updated when the DAnalysis will not necessarily be the
container of the representation. (through an adapter ?)

Temporary step in SessionResourceTracker.resolveAllVSMResources (must be
moved), to force resolution of controlled representations, before the
resolveAll which would add the repfile and then the vsm and environment
to semantic resources because of the reload done in addSemanticResource.

SiriusUtil.delete: add a way to retrieve the session in case of a
DRepresentation resource root


Bug: 493353
Change-Id: I98059dbe5cb7d397070eaba852fa689304f710ea
Signed-off-by: Maxime Porhel <maxime.porhel@obeo.fr>