[491895] DRAFT What this poc tries to prove:
* if it is possible to use the copy layout mechanism at the
semantic copy time
* if it is possible to use the layout information in the Canonical
Synchronizer (indeed, the GMF views and edit parts do not exist at the
semantic paste time, the information need to be retrieve during the GMF
views creation)
* if the SiriusLayoutDataManager can do the job with GMF Views as paste
target instead of EditParts. Until now, the paste layout is done on
diagram elements already displayed but it is not the case when applying
it during the GMF views creation (Canonical Synchronizer)

Limitations and remaining work:
* This poc showed a limitation with the key based on fragments. To
identify on which graphical element the layout should be apply, we use
by default the semantic uri fragment. But in the case of a copy paste
semantic, the fragment could be different.
Since this poc aims to prove the feasibility, an important work of
refactoring, mainly in the AbstractSiriusLayoutDataManager class is
remaining.

Bug: 491895
Change-Id: I6d50dfcbab9482b659ad5266906dad8d0e07f356
Signed-off-by: Florian Barbin <florian.barbin@obeo.fr>