[Draft] Distribute a selection of one should be applied on border nodes
If only one diagram element is selected, the distribute actions could be
applied to border node instead of being deactivated. In this draft, I
modified the Show Edit Part button to create a compound command that
distribute gap horizontally for the border node. It takes every border
node, but it should be improved to distribute each 4 borders
individually.

Change-Id: I0321e2cc8c4987fa7e44bfea2a0ad00a9b74818d
Signed-off-by: Steve Monnier <steve.monnier@obeo.fr>