[495240] Integrate rotatable node style.
This patchset provides a new custom style to define rotatable nodes. It
could be used for bordered node and node mappings.
To use it the specifier must specify a custom style for his node
mapping. Then it specifies a org.eclipse.sirius.rotatableImage extension
and set the same id than the one defined in the custom style.
Two possibilities exists:
* if only the north image is specified, this image will be rotated when
the node is moved.
* if all the images (north, south, east, west) are defined the depending
on the position of the node, the best image will be used to represent
the node.
 
In case of bordered node the image will change depending on which
container side the node is set. In case of node the first connection
angle is used to rotate the node image.

Bug: 495240
Change-Id: I743674bdc1f304562dc36ed5e30877c5eb97343f
Signed-off-by: Mélanie Bats <melanie.bats@obeo.fr>