[DRAFT][511783] Add no-op default operations to all newly created tools
This patchset is a possible alternative using a super class pour all
xxxItemProviderAdapter. This patchset is not completed because the work
has been done only for viewpoint.

Another alternative to study is to use ChildCreationExtender(on package
in genmodel)to extend ItemProviderFactory for a given package. It allows
to overload createChildParameter(Object feature, Object child). Then
overloading can be targeted on only the VSM packages.

Bug: 496019
Change-Id: Iaca0b2578b2320a1f6c6339816b550753ecd2c66
Signed-off-by: Laurent Fasani <laurent.fasani@obeo.fr>