[513309] add "diagram" variable and delay initialization execution
Many xxxTasks classes are relying on the capability to retrieve the
Session instance from a view to operate correctly, most notably every
CreateView task. 
These classes are failing to retrieve the session instance when the
operations are used within the context of the "diagramInitialization"
feature. Any CreateView used in this context will lead to a rollback of
the current transaction.

This commit delay the execution so that when the model operations are
executed the diagram is already attached to the session, as such the
xxxTasks will find the session instance. 
It also introduce a "diagram" variable at runtime and at specification
time so that using a CreateView operation is possible by specifying the
container view.

Bug: 513309
Change-Id: I88e64f70aa84db4b3acc9c78bafd6f75279f7495
Signed-off-by: Cedric Brun <cedric.brun@obeo.fr>