[513145] Do not require UI Thread to launch layout
The runnable executing the layout was called with asyncExec. In some
cases (for instance with an external change made in an XText editor)
this caused the diagramEditPart to arrange to have no rootEditPart
anymore and caused an NPE in GMF DeferredLayoutCommand.
The code is now called without guaranteeing that it is called on UI
Thread

Bug: 513145
Change-Id: I72accf4d373bfd65985a305f72270a5bdd489c9f
Signed-off-by: Steve Monnier <steve.monnier@obeo.fr>
Signed-off-by: Laurent Fasani <laurent.fasani@obeo.fr>