[DRAFT] Correctly set default size for container
If a container has a size defined in the VSM, with width and height of
SizeComputationContainerStyleDescription, it is not fully managed.

The GMF size after the diagram creation is {-1,-1} instead of the
defined size. This is caused by the AbstractCompositeLayoutProvider that
uses the ArrangeAllWithAutoSize. This last correctly considers the size
the first time in isDefaultSizeHasJustBeenSet() but for the following
calls to the same method, the result was different as the marker was
removed during the first call. This marker is now removed only during
the last call (in
ArrangeAllWithAutoSize.createSubCommands(CompoundCommand, Node, Point,
Rectangle)).

This commit is not finished. The problem is to detect how to remove the
adapter that says that the node must not be auto-size. If we uncomment
the line 135 of ArrangeAllWithAutoSize, the GMF size set is {-1, -1}.

Change-Id: I1dbb73d081153173ae64bb895adde5e303a9f1f8
Signed-off-by: Laurent Redor <laurent.redor@obeo.fr>