[DRAFT Test] Change default UIProject save to use session save
UIProject.save() now calls the session.save(IProgressMonitor) instead of
having SWTBot clicking on the save button that sometimes fail. This
behavior has been moved to a new method saveUsingUI().
The whole suite needs to be launched locally in order to check if this
change in behavior does not cause regressions.

Change-Id: I1eebddd7655154dde5cae4e2c291c6c9b7a6afdb
Signed-off-by: Steve Monnier <steve.monnier@obeo.fr>