[537111] Fix unwanted display of * in the Model Explorer view
The dirty status of a Session in the Model Explorer view depends on the
SAVE_WHEN_NO_EDITOR
preference. If the preference value is true (default behavior), the
status is hidden as soon
as a DialectEditor (diagram/table/tree) is opened as this status is
displayed in the editor tab.

Sirius 5.0 introduced the Aird editor and the ISiriusEditor API as
supertype of DialectEditor.

This commit generalizes the display of "*" (dirty status) behavior from
DialectEditor to all ISiriusEditor.

Bug: 537111
Change-Id: Id9288fd837b19d398425ac069bc95f9c6640b4f7
Signed-off-by: Maxime Porhel <maxime.porhel@obeo.fr>