WIP [472052] Workaround for selected table cells not visible on unix
This is a temporarary workaround. It removes the installation of the
listener that applies color for the selected line in Tables for MacOS
and Linux OS.
Even if the MacOS have the bigger issue with this listener (line content
not redable), its behavior is not satisfying with other systems (some
cases not readable in Linux for instance).

Bug: 472052
Change-Id: I2a748bcad73e23a9c7bd745140b0f77ce7ff6255
Signed-off-by: Mathieu Cartaud <mathieu.cartaud@obeo.fr>