DRAFT - Update tests using the properties view to change a value
Strangely the loading of the extension point for custom edition in the
properties view fails (once) if previously in the suite some test set
focus on an element value in the properties view. I edited some test to
update the value programmatically instead of using the UI. However,
ManualAirdModificationTest needs to use the UI as it is checking the
synchronisation with the editor. CellEditorExtensionTest may require to
be at the begining of the suite.

Bug: 551235
Change-Id: Ic71dd9d4d86c2556a8ab7deb9f25137cfc21cc7f
Signed-off-by: Steve Monnier <steve.monnier@obeo.fr>