[579818] Ignore visibility notifications for other diagrams
Assuming diagram1 and diagram2 are opened.
When applying the filter on diagram1, it will generate some visible
attribute changes on DDiagramElement.
This commit ignores these changes if the diagram of the DDiagramElement
is not the current diagram of the FilterListener.

So if n diagrams are opened we move from n+1 calls of
computeCompositeFilterApplications to 2 calls.

Bug: https://bugs.eclipse.org/bugs/show_bug.cgi?id=579818
Change-Id: Id25d482410c78b868b227e873050a5e5e53fa767
Signed-off-by: Laurent Fasani <laurent.fasani@obeo.fr>