DRAFT [573357] Cleanup variables before evaluation of tool precondition
This commit adds a cleanup steps before setting new variables values to
evaluate the precondition. This avoids to have side effect as explained
in the second remark of bug 573357 [1].

TODO: The impact of this cleanup must be checked in depth before merging
it. Indeed, there is a risk for other tools than the reconnect tools.

[1] https://bugs.eclipse.org/bugs/show_bug.cgi?id=573357#c0

Bug: 476523
Bug: 573357
Change-Id: I1295b1517c068d6ce91241726f2d31ac0a49525c
Signed-off-by: Laurent Redor <laurent.redor@obeo.fr>